requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-chroma
dashscope>=1.14.0
pyyaml>=6.0
pypdf>=4.0.0
chromadb>=0.4.22
python-dotenv>=1.0.0
pytest>=7.0.0
pymongo>=4.6.0
